(PE-5344) DRYed out authority-key-identifier Clojure function
This commit extracts some of the redundant map setup in the different
arities of the `authority-key-identifier` function into a helper
function, `create-authority-key-identifier`.  The helper function takes
advantage of the fact that the underlying Java method that interprets
the map was already able to handle the absence of a key in the map vs.
the presence of the key but with a null value as being equivalent.