(TK-121) Added CRL option to ca-cert only SSLContext
This commit adds the ability to include the CRL when creating an
SSLContext from a CA pem only, i.e., without a private key and
client/server certificate.  A `ca-cert-and-crl-pems->ssl-context`
Clojure function and `caCertAndCrlPemsToSSLContext` Java method were
introduced to support this.

This commit also includes a bit of refactoring intended to DRY up the
code a bit more and avoid including new exceptions related to CRL
handling on the throws clause for existing Java methods that don't take
a CRL argument.