(SERVER-119) Properly sign puppet extension values
Currently, this library signs all custom extensions as binary data wrapped
in a DER Octet String, which is the same way the Ruby Puppet Master does
it. According to RFC 5280, the values of all X.509 extensions must be a
valid DER encoded structure contained within a DER Octet String, so the
current method is technically incorrect.

The Bouncy Castle library assumes adherence to this standard, so trying
to exactly mimic the way MRI Puppet signs trusted facts has proven to be
very unwieldy and brittle.

Other extensions which contain a simple string as a value will use a
UTF8String type and then wrap it inside of a DER Octet String.