(SERVER-1545) Add helpers for getting subject from a certificate
This commit adds new helpers --
get-subject-from-x509-certificate in Clojure /
SSLUtils.getSubjectFromX509Certificate() in Java -- for extracting the
full subject from an X509Certificate object.  The helpers return the
subject in BCStyle since each of the related helpers already in
jvm-ssl-utils also use BCStyle.

This commit also changes each of the related helpers which implicitly
use BCStyle, by virtue of it being the default style, to explicitly use
BCStyle.  This is done to protect against the possibility of a future
default style change in BouncyCastle causing some of the helpers to
mangle string values if differing styles are used among helpers to
encode vs. decode values.