(PE-18417) Update to bouncycastle 1.55
The version of bouncycastle we are shipping is over 2 years old.
We've had a user report incompatibilities between it and recent
certificates generated by external tools.

This commit updates us to the most recent version, which contains
many bug fixes and features compared to the older version we were
shipping, but should be backward-compatible.