(SERVER-1315) Use key pair file when extracting ca-cert
The initial approach for extracting and verifying the CA certificate
used a single public key; unfortunately this doesn't mesh with what
Puppetserver expects. Changes made in SERVER-86 specifically tested that
certificate signing could happen without a ca public key. Since the
private keys we generate contain both the public and private key, we can
extract the public key from that pair and use that comparison. Switching
from a standalone public key to a private key with associated public key
should be more correct as we're doing comparisons with the content we'll
actually be using for signing.