(SERVER-1315) Collapse objectToKeyPair
The objectToKeyPair function was implemented to match the behavior used
for private keys in objectToPrivateKey, but generating key pairs is more
straight forward and doesn't really need a separate function. This
commit collapses the implementation of objectToKeyPair into
pemToKeyPairs, which is probably the only caller that this code will
have.