(SERVER-2205) Support loading CRL chains
This commit adds a general `objs->pem!` similar to the more specific
{crl,cert}->pem! functions so we may efficiently serialize chain/bundles
of PKI objects.

It also adds a pem->ca-crl to provide an analog to pem->ca-cert though
its usage in puppetserver does not allow us to require the ca cert to do
the degree of validation that pem->ca-cert does.

It pulls most of the `create-ca-extensions` from puppetserver
into this library to ease testing (creating chained CAs and their CRLs).

Finally, it modifies the existing "simple" namespace's gen-cert* and the
functions that call it to allow creating ca certs.

The last two changes, create-ca-extensions & gen-cert* are meant to aid
later integration tests by consumers of this library.