(PDB-4357) FIPS support
Move the bouncycastle jar to the dev profile. This will allow us
to specify the bouncycastle jars on the classpath at runtime like
we do for logging configuration. At installation time, we can install
the FIPS-compliant version of the jars, or the standard versions.

The FIPS version of the jar has a reduced number of classes compared to
the standard provider, but all the base functionality is there. Reworked
SSLUtils and ExtensionsUtils to use the classes that are provided in
both versions, and determine which provider based on availability on the
classpath.

Added additional build stages to Travis CI so that we test FIPS and
non-FIPS on both JDK 8 and JDK 11.