(SERVER-2019) Extract PublicKey from PEM when possible
When we added parsing of the CA cert from a potential bundle of certs we
retreived the CA's public key and validated the cert retrieved matched
the public key. This caused issues becaus the CA's public key is
optional as a distinct file.

The refactor in e9ba508 loaded the private key but assumed it could be
coerced into a keypair. Which makes sense in theory since the private
key contains all of the data needed to ceate a public key. However, the
PEMParser only returns the desired PEMKeyPair when parsing a PKCS#1
formatted file. This happens to be the file we use in all of our testing
and that the Puppet agent would create by default. However, it means
users that have private keys written in the newer PKCS#8 format, which
the PEMParser returns as PrivateKeyInfo objects, fail to properly load
the CA cert.

This moves us from calling pemToKeyPair() w/ the private key reader to
calling pemToPublicKey() and updating pemToPublicKey() to accept and
extract a PublicKey from a PEM reader containing any of:
SubjectPublicKeyInfo (as previously, a PKCS#1 formatted public key),
a PEMKeyPair (some PKCS#1 formatted private keys), or a PrivateKeyInfo
(PKCS#8 and some PKCS#1 formatted private keys).