(SERVER-2801) Catch IOException for malformed Netscape comment ext
This commit updates the exception we catch when parsing an unwrapped
Netscape comment cert extension from a legacy Puppet cert. Newer
versions of Bouncy Castle added an extra error case when attempting to
parse extension byte arrays, that our legacy certs hit, because the
extension values are not byte arrays but unwrapped strings. Previously,
we would hit an EOFException, but now it throws a generic IOException
before proceeding to the check we used to fail. IOException is a
superclass of EOFException, so by catching the former, we are also
backwards compatible.