(maint) Add optional `nextUpdate` arg to `generateCRL`
This commit adds a version of the `generateCRL` method that takes a `nextUpdate`
argument. Normally, CRLs are generated with a `nextUpdate` time of five years in
the future. This change allows us to generate CRLs with any `nextUpdate` time,
including times in the past, which is useful for testing expired CRLs.