(maint) Allow specifying your own key pair when generating test files
This commit adds an optional `root-key-pair` parameter to the
`generate-cert-chain-with-crls` function. This allows us to specify our own key
pair when generating test certs and crls, rather than generating the key pair
as part of the function. This is useful for testing additional operations on
certs and CRLS, e.g. revocation, that require access to the key pair -
previously, the keys used to generate the test files were just lost afterwards.