(SERVER-2550) Add `crlNumber` and `extensions` params to `generateCRL`
This commit adds a `crlNumber` param and an `extensions` param to the
`generateCRL` method. Normally, CRLs are generated with a CRL Number of 0 and
only the AuthorityKeyIdentifier and CRLNumber extensions. This change allows us
to generate CRLs with any CRL Number and any additional extensions, which is
useful for testing CRL updates.