(maint) Ensure compatibility with bc-fips API
Versions of bouncy castle past 1.63 introduced additional `equals`
methods to the `ASN1ObjectIdentifier` class, with more specific type
signatures. However, these methods to not exist in FIPS BC. This commit
adds some casts to `Object` to places where we call this method, to
ensure we always get the more general API that exists in both versions.

Note that this caused an issue when ssl-utils class files compiled with
newer non-FIPS BC were used when running with FIPS inside Puppetserver.
The compiled files expected the new APIs, which were not available when
the FIPS jar was used at runtime.