(maint) Allow CA extensions to include auth key id from issuer public key
Previously, we only allowed generating CA cert extensions with serial
and issuer values for the authority key identifier. This is not valid
with some libraries (notably Ruby's openssl), which requires the auth
key id to include the key ID from the issuer's public key. This adds an
arity to the function that generates CA extensions, to allow generating
the more correct form.