(maint) Add `write-bundle-for-manual-validation` testutil
This commit adds a test utility for generating the necessary certs, CRLs, and
keys to be used with the `puppetserver ca import` command. This is useful for
manual testing, which is why it doesn't write files to this repo's
`test-resources` directory by default. It also includes 'newer' and 'newest'
CRLs, which is useful for testing the CA endpoint for updating CRLs.