(SERVER-3016) Filter out EC params when parsing private keys
When OpenSSL generates an elliptic curve key, it writes the EC params in
a separate block above the key in the output file. When we parse this
file, we load both objects in the file, the EC params, and the key. But
we only want to return the key and ignore the params, so this commit
adds a filter to the reader function, so that EC param objects are
ignored.