(SERVER-2509) Prevent duplicate entries in Puppet's CA CRL
This commit contain a function that compare the list of serials to be add to the revoked list against the already existed list of serials of revoked cert.  The function will return a new list of serials to be added to the revoked list without any serial that already existed.