(SERVER-2509) Add conditional check for revoking CRL
This commit includes a conditional check to the function
'revoke-multiple'. The conditional check's purpose is
to prevent modifying the CRL if there is no 'new' serial
number to be added to the revoked list. If there is no
'new' serial number, simply return the supplied CRL.
Otherwise, proceed with adding the serial number to the
revoked list and update the CRL.