(PE-36454) add support for adding attributes to csrs, fix issues
This adds support in the "simple" namespace for adding attributes
to CSRs. It adds a test to demonstrate the behavior.

It also fixes several problems, including a typo in the "get-keys" function
which was associating a name with the wrong key, as well as removing use
of a deprecated conversion routine.

Some require organization was done as well.