(maint) prevent stack overflows
Changes in BouncyCastle removed the `toObject` function as part of an ASN1Primitive, and
a fix was put in that change the code based on the interace provided. Unfortunately this
fix could result in an stack overflow for some certificates. This removes the
opportunity for that, and uses the default encoding case because ASN1Tagged items are
already ASN1Primitive items.